net/netns: move SOCKS dialing to netns for now
This lets control & logs also use SOCKS dials.
